Use the chat() react ai hook  correctly
Remove isLoading state, useState hook
