+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@ai-sdk/openai": "^1.2.5",
     "@ai-sdk/provider": "^1.0.11",
     "@ai-sdk/provider-utils": "^2.1.13",
     "@ai-sdk/ui-utils": "^1.1.19",
@@ -17,7 +18,7 @@
     "@radix-ui/react-slider": "^1.2.3",
     "@radix-ui/react-slot": "^1.1.2",
     "@vercel/analytics": "^1.5.0",
-    "ai": "^3.0.19",
+    "ai": "^3.4.33",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.479.0",
@@ -33,7 +34,7 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
+    "@types/node": "20.17.24",
     "@types/react": "19.0.10",
     "@types/react-dom": "^19",
     "eslint": "^9",
